Add initial mnemonic support
Introduce the assembly translation.
This is accomplished leveraging a definitions file, where every instruction
is placed and described (opcode, prefixes etc) along with an handler.

The definition file is parsed by a Python script (Gemini generated almost
entirely) in order to produce a file header, "generated_opcodes.h".

The file header contains the call to the handler.

This is just a first attempt, the python script may be changed in future.